Moved KANJI_MAP to global variable
